feat: sequence Comparable 구현, (자료형 Long -> Integer로 변경)

+++ src/main/java/kr/co/blossomedu/vocatest/chapters/service/dto/ChapterCreateRequest.java
@@ -12,11 +12,11 @@ import lombok.ToString;
 @ToString
 @NoArgsConstructor
 public class ChapterCreateRequest {
-    private Long sequence;
+    private Integer sequence;
     private String name;
     private Long bookId;
 
-    public ChapterCreateRequest(final Long sequence, final String name, final Long bookId) {
+    public ChapterCreateRequest(final Integer sequence, final String name, final Long bookId) {
         this.sequence = sequence;
         this.name = name;
         this.bookId = bookId;
